since the flux capacitor stencil moved into d_sw, moving the Translate class there as well
