+++ tests/translate/__init__.py
@@ -20,14 +20,18 @@ from .translate_corners import (
 )
 from .translate_cubedtolatlon import TranslateCubedToLatLon
 from .translate_d2a2c_vect import TranslateD2A2C_Vect
-from .translate_d_sw import TranslateD_SW, TranslateUbKE, TranslateVbKE
+from .translate_d_sw import (
+    TranslateD_SW,
+    TranslateFluxCapacitor,
+    TranslateUbKE,
+    TranslateVbKE,
+)
 from .translate_del2cubed import TranslateDel2Cubed
 from .translate_del6vtflux import TranslateDel6VtFlux
 from .translate_delnflux import TranslateDelnFlux, TranslateDelnFlux_2
 from .translate_divergencedamping import TranslateDivergenceDamping
 from .translate_dyncore import TranslateDynCore, TranslatePGradC
 from .translate_fillz import TranslateFillz
-from .translate_fluxcapacitor import TranslateFluxCapacitor
 from .translate_fvsetup import TranslateFVSetup
 from .translate_fvsubgridz import TranslateFVSubgridZ
 from .translate_fvtp2d import TranslateFvTp2d, TranslateFvTp2d_2
